content_automation/requirements.txt
asyncpg==0.29.0
aiohttp==3.9.1
openai==1.51.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6

# Optional: For local video processing
# ffmpeg-python==0.2.0

# Optional: For advanced analytics
# pandas==2.1.4
# matplotlib==3.8.2